Exception handling

+++ src/main/java/com/carlosguilherme/dscommerce/services/ProductService.java
@@ -1,16 +1,21 @@
 package com.carlosguilherme.dscommerce.services;
 
-import java.util.Optional;
-
+import org.hibernate.tool.schema.internal.StandardAuxiliaryDatabaseObjectExporter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.carlosguilherme.dscommerce.dto.ProductDTO;
 import com.carlosguilherme.dscommerce.entities.Product;
 import com.carlosguilherme.dscommerce.repositories.ProductRepository;
+import com.carlosguilherme.dscommerce.services.execptions.DatabaseExecption;
+import com.carlosguilherme.dscommerce.services.execptions.ResourceNotFoundExecption;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 
@@ -22,10 +27,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
-		ProductDTO dto = new ProductDTO(product);
-		return dto;
+		Product product = repository.findById(id).orElseThrow(()-> new ResourceNotFoundExecption("Recurso não encontrado."));
+		return new ProductDTO(product);
 		
 	}
 	
@@ -48,17 +51,32 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoEntity(dto, entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoEntity(dto, entity);
+			
+			entity = repository.save(entity);
+			
+			return new ProductDTO(entity);
+			
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundExecption("Recurso não encontrado");
+		}
 		
-		entity = repository.save(entity);
-		
-		return new ProductDTO(entity);
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundExecption("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseExecption("Falha de integridade referencial");
+		}
+		
 		
 	}
